+++ spec/lib/system_checker_spec.rb
@@ -17,11 +17,11 @@ module Decidim::DecidimAwesome
     end
 
     it "has 13 modified files in core" do
-      expect(subject.overrides["decidim-core"].files.length).to eq(13)
+      expect(subject.overrides["decidim-core"].files.length).to eq(15)
     end
 
-    it "has 18 modified files in proposals" do
-      expect(subject.overrides["decidim-proposals"].files.length).to eq(18)
+    it "has 14 modified files in proposals" do
+      expect(subject.overrides["decidim-proposals"].files.length).to eq(14)
     end
 
     context "when file" do